Refactor frontend JS into feature modules
