+++ frontend/js/modules/login/auth.js
@@ -1,4 +1,4 @@
-import { API_URL } from './config.js';
+import { API_URL } from '../../config.js';
 
 const STORAGE_KEY = 'reportesOBM.user';
